Refactor repository structure and implement kebab-case naming convention
- Established a clear directory structure for components, hooks, and utilities.
- Renamed existing files to adhere to the kebab-case naming convention.
- Moved utility functions and constants to a new `lib` directory.
- Updated import paths throughout the codebase to reflect the new structure.
- Removed unused files and consolidated related functionality into appropriate directories.
- Ensured all tests and components are functioning correctly after restructuring.
