+++ src/components/modal.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { StateManager } from 'cotton-box';
 import { ModalRootProps } from '@decky/ui';
 
-import { State, GlobalContext } from './globalState';
+import { State, GlobalContext } from '../hooks/global-state';
 
 interface ModalContext extends ModalRootProps {
   value: StateManager<State>;
