+++ src/lib/geometry.ts
